requirements.txt
django
django-autocomplete-light
django-bootstrap4
celery
django-model-utils
redis
